Move package.json to server folder for docker compatibility
In order to conveniently containerize the server, all files that need to be added to the
docker image must be in the same directory as the dockerfile. As the db container does
not use node packages, it makes sense to move the package.json file to the server folder
and separate the containers further.