Separate logic for db schema setup and csv data loading
Previously there was an init.sql file that loaded the schema into the database and copied the
data from the CSV files then loaded it into the database. Now schema.sql contains the schema
logic and tl.sql (transfer, load) copies the data from the CSV files and loads it into the datbase.