Set up link to database container from server and persistent db storage
A link with the alias "database" has been configured, connecting the database to the server
container. Volumes for persisting database data across database container restarts and a volume
for setting up a schema and loading CSV data into the database has been configured. The
latter volume is only run when no database is created yet (initialization)