Change model functions to use clients instead of querying pool directly
Previously each model function queried the database by using `pool.query()`, which meant that
queries were made one after the other in a serial fashion. This was alright for testing but
is not suitable for a frequent requests. Switching to creating clients as instances of
`pool.connect()` and running `.query` on each client, multiple clients could be checked out
and the server would be better equipped for more frequent requests.