Optimize getProductInfoById to not create unnecessary clients
Previously a client was created and released when querying the database for the features
of the product associated with the passed in ID and also when querying for the product
from the products table. These two query functions have been combined so that only one
client is created for both getting the product and it's features. This should optimize the
use of clients for when there are a high frequency of requests.