Swap config file for docker-compose for easier deployment
Docker-compose makes the starting up of containers easier since there is no need to build a
local image then run that image as docker-compose can be configured to conveniently pull
and run an image from dockerhub. It also allows port mapping to be set up so it does
not have to be done in the command line. The config file has been removed as docker-compose
allows the setting of environment variables.