Replace getSkus, getPhotos, and getStyles with getStyles
The queries for getting skus and photos from a style, as well as getting styles from a product, have been combined for optimization so the separate functions for each query are no longer
necessary. As a result, server latency for the styles endpoint is down as less queries are being made and subsequently less clients being taken up from the pool.