Implement redis caching for styles endpoint
If the styles endpoint recieves a request for an item that has already been cached, it will query the redis cache rather than the database for that information. If the request is for data
that has not been cached, it will query the postgres database and then cache that item.